feat: Add support for ignoring function names starting with _
This commit adds a new feature to the `extract_docs.sh` script. Now, when processing files, if a function name starts with an underscore (_), it will be ignored and not added to the man page content. This change improves the accuracy of the generated documentation by excluding functions that are intended for internal use only.