Add snapc function
Add snapc function

This commit adds a new function called `snapc` to the `.functions` file. The `snapc` function is used to clean up old snap versions. It checks if snap is installed and if the user is running as root before removing old revisions of snaps using the `snap remove` command. This function helps in managing and maintaining snap packages efficiently.